fixed alignment wooo

+++ core/src/com/pear/pudding/model/Board.java
@@ -19,10 +19,10 @@ public class Board extends Deck {
         setWidth(width);
         setHeight(height);
         setLocation(Location.BOARD);
-        setSlotWidth(width / NUMBER_OF_BOARD_SLOTS);
+        setSlotWidth(width / (float) NUMBER_OF_BOARD_SLOTS);
 
         int middleSlotIndex = NUMBER_OF_BOARD_SLOTS / 2;
-        for (int i = 0; i < NUMBER_OF_BOARD_SLOTS; i++) {
+        for (int i = 0; i < (float) NUMBER_OF_BOARD_SLOTS; i++) {
             Side slotSide = i < middleSlotIndex ? Side.LEFT : (i > middleSlotIndex ? Side.RIGHT : Side.MIDDLE);
             var slot = new Slot(getX() + i * getSlotWidth(), getY(), getSlotWidth(), getHeight(), i);
             slot.setSide(slotSide);
